ALELUYA

+++ src/property/property.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { CreatePropertyDto } from './dto/create-property.dto';
 import { UpdatePropertyDto } from './dto/update-property.dto';
-
+import { ForbiddenException } from '@nestjs/common';
 import { PrismaService } from 'src/prisma.service';
 import { Property } from '@prisma/client';
 @Injectable()
@@ -24,6 +24,14 @@ export class PropertyService {
     return this.prisma.property.findMany();
   }
 
+  findByUserId(userId: number): Promise<Property[]> {
+    return this.prisma.property.findMany({ 
+      where: {
+        userId,
+      },
+    });
+  }
+
   findPropertyByID(id: number): Promise<Property | null> {
     return this.prisma.property.findUnique({
       where: {
@@ -32,20 +40,28 @@ export class PropertyService {
     });
   }
 
-  update(id: number, data: UpdatePropertyDto): Promise<Property> {
+  async update(id: number, data: UpdatePropertyDto, userId: number) {
+    const property = await this.prisma.property.findUnique({ where: { id } });
+
+    if (!property || property.userId !== userId) {
+      throw new ForbiddenException('No puedes editar esta propiedad');
+    }
+
     return this.prisma.property.update({
-      where: {
-        id,
-      },
+      where: { id },
       data,
     });
   }
 
-  remove(id: number): Promise<Property> {
+  async remove(id: number, userId: number) {
+    const property = await this.prisma.property.findUnique({ where: { id } });
+
+    if (!property || property.userId !== userId) {
+      throw new ForbiddenException('No puedes eliminar esta propiedad');
+    }
+
     return this.prisma.property.delete({
-      where: {
-        id,
-      },
+      where: { id },
     });
   }
 }
